#36 reorganizes databases directory into a unfied directory, closes (#36)
