+++ backend/app/database/mongodb/repository.py
@@ -9,7 +9,7 @@ from pydantic import BaseModel
 
 from pymongo.collection import Collection
 
-from app.db.client import get_database
+from . import get_database
 
 # Configure logging
 logging.basicConfig(level=logging.INFO)
