+++ backend/app/database/postgre/__init__.py
@@ -1,0 +1,2 @@
+from .config import DATABASE_URL
+from .database import engine, Base, get_db
