+++ backend/app/models/blacklist_token.py
@@ -3,7 +3,7 @@ from sqlalchemy import Column, String, func, DateTime
 from sqlalchemy.dialects.postgresql import UUID
 from datetime import datetime, timezone
 
-from app.core.database import Base
+from app.database.postgre import Base
 
 class BlackListToken(Base):
     __tablename__ = "blacklisted_tokens"
