+++ backend/app/models/otp.py
@@ -1,6 +1,6 @@
 from sqlalchemy import Column, String, Integer, Boolean, DateTime
 from sqlalchemy.sql import func
-from app.core.database import Base
+from app.database.postgre import Base
 
 class OTP(Base):
     __tablename__ = "otp"
