+++ backend/app/models/scans.py
@@ -3,7 +3,7 @@ from datetime import datetime, timezone
 from sqlalchemy import Column, String, ForeignKey, DateTime, UniqueConstraint, Index
 from sqlalchemy.dialects.postgresql import UUID
 from sqlalchemy.orm import relationship
-from app.core.database import Base
+from app.database.postgre import Base
 
 class Scans(Base):
     __tablename__ = "scans"
