+++ backend/app/models/subscriptions.py
@@ -3,7 +3,7 @@ from datetime import datetime, timezone
 from sqlalchemy import Column, String, Integer, DateTime, Enum
 from sqlalchemy.dialects.postgresql import UUID
 from sqlalchemy.orm import relationship
-from app.core.database import Base
+from app.database.postgre import Base
 
 from .enums import AnalysisLevel, ScanDepth
