+++ backend/app/models/users.py
@@ -4,7 +4,7 @@ from sqlalchemy.dialects.postgresql import UUID
 from sqlalchemy.orm import relationship
 from datetime import datetime, timezone, timedelta
 
-from app.core.database import Base
+from app.database.postgre import Base
 
 class Users(Base):
     __tablename__ = "users"
