+++ backend/app/modules/iam_enumeration.py
@@ -4,7 +4,7 @@ from datetime import datetime
 
 from app.modules.iam.enumerator import IAMEnumerator
 from app.modules.iam.exceptions import IAMEnumerationError, CredentialsError
-from app.db import save_snapshot
+from app.database.mongodb import save_snapshot
 from app.monitoring.metrics import enumeration_duration_seconds, enumeration_errors_total
 
 # Configure logging
