+++ backend/app/modules/lambda_enumeration.py
@@ -3,7 +3,7 @@ import time
 from datetime import datetime, timezone
 
 # Import the Lambda enumeration system
-from app.db import save_snapshot
+from app.database.mongodb import save_snapshot
 from app.modules.aws_lambda.enumerator import LambdaEnumerator
 from app.modules.aws_lambda.exceptions import LambdaEnumerationError, LambdaCredentialsError
 from app.monitoring.metrics import enumeration_errors_total
