+++ backend/app/modules/s3_enumeration.py
@@ -6,7 +6,7 @@ from datetime import datetime
 from app.modules.s3.enumerator import S3Enumerator
 from app.modules.s3.exceptions import S3EnumerationError, S3CredentialsError
 
-from app.db.repository import save_snapshot
+from app.database.mongodb import save_snapshot
 from app.monitoring.metrics import enumeration_duration_seconds, enumeration_errors_total
 
 # Configure logging
