+++ backend/app/service/genai_service.py
@@ -1,10 +1,8 @@
 import logging
 from sqlalchemy.orm import Session
-from dataclasses import asdict
-from datetime import datetime
 from typing import Optional
 
-from app.db import store_result, get_by_scan_id
+from app.database.mongodb import store_result, get_by_scan_id
 from app.schemas import CurrentUser, AttackPath
 from genai.data_preprocessor import DataPreprocessor
 from genai.risk_analyzer import RiskAnalyzer
