+++ backend/app/service/scan_service.py
@@ -8,7 +8,7 @@ from app.repositories import ScanRepository, EnvironmentRepository, UserReposito
 from app.schemas import CurrentUser
 from app.modules import Enumerator
 from .genai_service import GenAIService
-from app.db.repository import get_by_scan_id
+from app.database.mongodb import get_by_scan_id
 from app.mappers.scan_results_mapper import ScanResultMapper
 from app.utils import collections
 from app.monitoring.metrics import (
