+++ backend/app/utils/dependency/auth_dependencies.py
@@ -7,7 +7,7 @@ from datetime import datetime, timezone
 from app.schemas import CurrentUser
 from app.models.enums import ScanDepth, AnalysisLevel
 from app.utils import decode_access_token
-from app.core.database import get_db
+from app.database.postgre import get_db
 from app.repositories import UserRepository, TokenBlacklistRepository
 
 # HTTP Bearer token scheme
