+++ backend/app/utils/dependency/service_dependencies.py
@@ -1,7 +1,7 @@
 from __future__ import annotations
 
 from sqlalchemy.orm import Session
-from app.core.database import get_db
+from app.database.postgre import get_db
 from fastapi import Depends
 
 # Dependency to get ScanService instance
