+++ backend/logs/app.log
@@ -3333,3 +3333,6 @@
 {"timestamp": "2026-03-19T10:48:22.697195Z", "level": "INFO", "logger": "app", "message": "Logging initialized", "service": "backend-app", "hostname": "192.168.10.4", "file": "/Users/yarkhan/Tech/AWSPentest/backend/app/logging/logging_config.py", "line": 102, "function": "setup_logging"}
 {"timestamp": "2026-03-19T11:02:16.463883Z", "level": "INFO", "logger": "app", "message": "Logging initialized", "service": "backend-app", "hostname": "192.168.10.4", "file": "/Users/yarkhan/Tech/AWSPentest/backend/app/logging/logging_config.py", "line": 102, "function": "setup_logging"}
 {"timestamp": "2026-03-19T11:02:32.114007Z", "level": "INFO", "logger": "app", "message": "Logging initialized", "service": "backend-app", "hostname": "192.168.10.4", "file": "/Users/yarkhan/Tech/AWSPentest/backend/app/logging/logging_config.py", "line": 102, "function": "setup_logging"}
+{"timestamp": "2026-03-23T07:04:30.404691Z", "level": "INFO", "logger": "app", "message": "Logging initialized", "service": "backend-app", "hostname": "192.168.10.4", "file": "/Users/yarkhan/Tech/AWSPentest/backend/app/logging/logging_config.py", "line": 102, "function": "setup_logging"}
+{"timestamp": "2026-03-23T08:07:10.595245Z", "level": "INFO", "logger": "access", "message": "HTTP request", "service": "backend-app", "hostname": "192.168.10.4", "file": "/Users/yarkhan/Tech/AWSPentest/backend/app/logging/logging_middleware.py", "line": 44, "function": "dispatch", "correlation_id": "b4eb0fd3-295d-4b59-a451-8431693ee306", "method": "GET", "path": "/genai/75c5c795-5cb3-4962-a7f7-cf85f5ac203a/", "status_code": 307, "duration_ms": 0.39, "is_slow": false, "client_ip": "127.0.0.1", "user_id": null}
+{"timestamp": "2026-03-23T08:07:10.600339Z", "level": "WARNING", "logger": "access", "message": "HTTP 4xx error", "service": "backend-app", "hostname": "192.168.10.4", "file": "/Users/yarkhan/Tech/AWSPentest/backend/app/logging/logging_middleware.py", "line": 40, "function": "dispatch", "correlation_id": "125d427a-3148-4fb6-b841-5ea6384d2941", "method": "GET", "path": "/genai/75c5c795-5cb3-4962-a7f7-cf85f5ac203a", "status_code": 401, "duration_ms": 2.41, "is_slow": false, "client_ip": "127.0.0.1", "user_id": null}
