+++ backend/main.py
@@ -6,7 +6,7 @@ from prometheus_fastapi_instrumentator import Instrumentator
 
 from app.logging import lifespan
 from app.logging import RequestLoggingMiddleware
-from app.core.database import engine, Base
+from app.database.postgre import engine, Base
 from app.routers import environment_router, scan_router, genai_router, auth_router
 
 # Initialize FastAPI app
